seasonal ratings (#138)
This commit defines seasonal ratings in what we expect to be the final form for 1.0. It provides a means to define season day ratings and seasonal night ratings. It defines a seasonal ratings snapshot media for `GET` operations and two seasonal ratings proposal media types for `PATCH` operations. One of the proposal media types is a "slim" representation that makes the most common case of using the same units for all individual resource proposals much more space efficient.